+++ ml-agents/mlagents/trainers/tests/tensorflow/test_models.py
@@ -2,14 +2,13 @@ import pytest
 
 from mlagents.trainers.tf.models import ModelUtils
 from mlagents.tf_utils import tf
-from mlagents_envs.base_env import BehaviorSpec, ActionType
+from mlagents_envs.base_env import BehaviorSpec, ActionSpec
 
 
 def create_behavior_spec(num_visual, num_vector, vector_size):
     behavior_spec = BehaviorSpec(
         [(84, 84, 3)] * int(num_visual) + [(vector_size,)] * int(num_vector),
-        ActionType.DISCRETE,
-        (1,),
+        ActionSpec.create_discrete((1,)),
     )
     return behavior_spec