+++ ml-agents/mlagents/trainers/tests/test_rl_trainer.py
@@ -8,6 +8,8 @@ from mlagents.trainers.tests.test_buffer import construct_fake_buffer
 from mlagents.trainers.agent_processor import AgentManagerQueue
 from mlagents.trainers.settings import TrainerSettings, FrameworkType
 
+from mlagents_envs.base_env import ActionSpec
+
 
 # Add concrete implementations of abstract methods
 class FakeTrainer(RLTrainer):
@@ -92,7 +94,7 @@ def test_advance(mocked_clear_update_buffer, mocked_save_model):
         length=time_horizon,
         observation_shapes=[(1,)],
         max_step_complete=True,
-        action_space=[2],
+        action_spec=ActionSpec.create_discrete((2,)),
     )
     trajectory_queue.put(trajectory)
 
@@ -144,7 +146,7 @@ def test_summary_checkpoint(mock_add_checkpoint, mock_write_summary, framework):
         length=time_horizon,
         observation_shapes=[(1,)],
         max_step_complete=True,
-        action_space=[2],
+        action_spec=ActionSpec.create_discrete((2,)),
     )
     # Check that we can turn off the trainer and that the buffer is cleared
     num_trajectories = 5